requirements.txt
python-dotenv
streamlit
langchain
langchain-classic
langchain-openai
langchain-chroma
langchain-community
langchain-text-splitters
chromadb
pypdf
